Adds PlayerPatcher and a hack to fix a Vanilla bug.
Removes ID checks when loading items; avoids invalid item IDs resulting in the player file being treated as corrupt (because it fails to read the correct number of bytes, since it skips stack/prefix/favourite).

Should make it easier to overload the save/load process, and if the bug gets fixed in game itself, the hack can be removed in the future.