Adds an UpdateMusic() hook.
The hook executes immediately after Main.curMusic gets set to Main.instance.newMusic; as a result, you can change Main.curMusic to set the currently-playing music.

Example mod altered to perform a similar proximity check with the PizzaBoss as vanilla does with other bosses, setting the music to 25 (Boss 5/Queen Bee theme).

The hook is implemented by creating a dummy method "UpdateMusicHook" inside Terraria.Main, adding a virtual method call to it inside UpdateMusic(), and then overriding it with a call to the new hook in TMain. It's very convoluted, but it works. Somewhat inefficient to iterate through the NPC array again looking for the boss, but trying to insert into the middle of the NPC check would make the implementation of the hook massively more complicated.